// server/src/enums/StatusCode.enum.ts
export enum StatusCode {
    OK = 200,
    CREATED = 201,
    ACCEPTED = 202,
    BAD_REQUEST = 400,
    UNAUTHORIZED = 401,
    FORBIDDEN = 403,
    NOT_FOUND = 404,
    INTERNAL_SERVER_ERROR = 500
}

export function StatusCodeGenerate(statusCode: StatusCode) {
    return {
        status: StatusCode[statusCode],
        code: statusCode
    }
}